test: split manager executor coverage
